api_gateway/api_gateway.py
# api_gateway.py
from flask import Flask, request
from werkzeug.exceptions import HTTPException
import logging
import yaml

app = Flask(__name__)

# Load configuration files
with open('config/routing.yaml', 'r') as file:
    routing_config = yaml.safe_load(file)

with open('config/authentication.yaml', 'r') as file:
    authentication_config = yaml.safe_load(file)

with open('config/rate_limiting.yaml', 'r') as file:
    rate_limiting_config = yaml.safe_load(file)

# Apply middleware
from middleware import logging, error_handling

app.register_blueprint(logging.logging_bp)
app.register_error_handler(HTTPException, error_handling.handle_http_exception)

# Define route for each service based on routing configuration
for route, service_url in routing_config.items():
    @app.route(route, methods=['GET', 'POST', 'PUT', 'DELETE'])
    def proxy(route=route, service_url=service_url):
        # Implement logic to forward requests to the appropriate service
        # This can involve authentication, rate limiting, etc.
        pass

if __name__ == '__main__':
    app.run(debug=True)
