api_gateway/requirements.txt
Flask==2.1.0
PyYAML==6.0